ai/requirements.txt
openai>=1.40
pydantic>=2.7
pillow>=10.3
